+++ src/util.ts
@@ -1,6 +1,7 @@
 import type { Subscriptions } from '@/types'
+import type { Webview } from 'vscode'
 
-import * as vscode from 'vscode'
+import { Uri, commands } from 'vscode'
 
 
 export const initCommand = (
@@ -9,6 +10,23 @@ export const initCommand = (
     command: (...args: any[]) => any
 ) => {
     subscriptions.push(
-        vscode.commands.registerCommand(commandId, command)
+        commands.registerCommand(commandId, command)
     )
+}
+
+export const getNonce = () => {
+    let text = ""
+    const possible = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789"
+    for (let i = 0; i < 32; i++) {
+        text += possible.charAt(Math.floor(Math.random() * possible.length))
+    }
+    return text
+}
+
+export const getUri = (
+    webview: Webview, 
+    extensionUri: Uri, 
+    pathList: string[]
+) => {
+    return webview.asWebviewUri(Uri.joinPath(extensionUri, ...pathList))
 }