+++ webpack.config.ts
@@ -1,13 +1,18 @@
 import * as path from 'path'
 import * as webpack from 'webpack'
 import { TsconfigPathsPlugin } from 'tsconfig-paths-webpack-plugin'
+import * as CopyPlugin from 'copy-webpack-plugin'
 
 const config: webpack.Configuration = {
   target: 'node',
-  entry: './src/extension.ts',
+  entry: {
+    extension: './src/extension.ts',
+    webview: './src/webview/main.ts',
+    vscodeUI: './node_modules/@vscode-elements/elements/dist/vscode-tree/vscode-tree.js'
+  },
   output: {
     path: path.resolve(__dirname, 'dist'),
-    filename: 'extension.js',
+    filename: '[name].js',
     libraryTarget: 'commonjs2',
     devtoolModuleFilenameTemplate: '../[resource-path]'
   },
@@ -35,7 +40,24 @@ const config: webpack.Configuration = {
         ]
       }
     ]
-  }
+  },
+  plugins: [
+    new CopyPlugin({
+      patterns: [
+        { 
+          from: 'node_modules/@vscode/codicons/dist', 
+          to: './',
+          filter: (path) => (/.*(\.ttf|\.css)/gm).test(path)
+        },
+        { 
+          from: 'media/instructionsList.css', 
+          to: './',
+          filter: (path) => (/.*(\.ttf|\.css)/gm).test(path)
+        },
+        // { from: 'node_modules/@vscode-elements/elements/dist/bundled.js', to: 'vscodeUI.js'}
+      ]
+    })
+  ]
 }
 
 export default config